requirements.txt
# Discord Self-Bot Library
discord.py-self>=1.9.0,<2.0.0

# Async HTTP client (compatible with discord.py-self)
# discord.py-self requires aiohttp<3.8.0
aiohttp>=3.6.0,<3.8.0

# Data handling
python-dateutil>=2.8.0,<3.0.0

# Image Processing (OpenCV + NumPy)
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0

# OpenAI for GPT-4 Vision (primary extraction method)
openai>=1.0.0,<2.0.0

# PaddleOCR for hybrid architecture (98%+ accuracy)
# Install CPU version first, upgrade to GPU if needed
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Optional: For enhanced console output
# colorama>=0.4.6
# rich>=13.0.0
